db/admin

+++ base/views.py
@@ -1,7 +1,11 @@
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.messages.storage.base import Message
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
 from django.contrib import messages
-from .models import Room, Topic
+from django.contrib.auth.decorators import login_required
+from .models import Room, Topic, Message
 from django.db.models import Q
 from django.contrib.auth.models import User
 from .forms import RoomForm
@@ -16,22 +20,46 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method =='POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
             user = User.objects.get(username=username)
         except:
-            messages.error(request, 'User does not exist')
+            messages.error(request, 'User or Password does not exist')
         user = authenticate(request, username=username, password=password)
         if user is not None:
             login(request, user)
             return redirect('home')
+        else:
+            messages.error(request, 'Username or Password does not exist')
 
-    context = {}
+    context = {'page':page,}
     return render(request, 'base/login_register.html', context)
 
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    page = 'register'
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error has occurred during registration")
+    return render(request, 'base/login_register.html', {'form':form,})
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
 
@@ -48,9 +76,20 @@ def home(request):
     return render(request, 'base/home.html', context)
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room':room}
-    return render(request, 'base/room.html', context)
+    room_messages=room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
 
+    context = {'room':room, 'room_messages':room_messages, 'participants':participants,}
+    return render(request, 'base/room.html', context)
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -61,10 +100,13 @@ def createRoom(request):
 
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -74,10 +116,22 @@ def updateRoom(request, pk):
 
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!')
     if request.method =='POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj':room})
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    messages = Message.objects.get(id=pk)
+    if request.user != messages.user:
+        return HttpResponse('You are not allowed here!')
+    if request.method =='POST':
+        messages.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj':messages})
